refactor(tir): semantic — analysis-framework-inline (-8 net lines)
Inline the single-instantiation analysis.rs dataflow framework into throw_inference.rs. Verified: unit + full baml_tests snapshot suite green, zero snapshot drift, workspace clippy clean, adversarial diff review passed.
